Switch to death data by date of death.
The reporting of deaths was all over the place over Christmas, resulting in unrepresentative graphs of weekly changes. Using the date of death improves the quality of the graph. It makes it a bit less live, as you have to ignore the most recent few days that are incomplete. On https://coronavirus.data.gov.uk/details/deaths, it says the most recent 5 days are incomplete, but this information isn't included in the JSON, so I manually ignore the most recent 5 days in covid-uk.js.

I haven't made the same change for the case data, as that doesn't seem to have the same data quality issues, so I'll keep the "by date reported" to keep liveness.